\name{betablok}
\docType{data}
\alias{betablok}
\title{Beta-blockers for myocardial infarction}
\description{
  A 22-centre clinical trial of beta-blockers for
reducing mortality after myocardial infarction, described by Yusuf \emph{et. al.}
(1984). The important issue is the generalizability of the treatment
effect across different patient populations.
}
\usage{betablok}
\format{ A dataframe with 44 obs. and 4 variables: 
  \tabular{rll}{
    [,1]\tab r \tab integer  \cr
    [,2]\tab n \tab integer \cr
    [,3]\tab centre\tab int \cr
    [,4]\tab treat \tab factor w /2 levels "C","T"\cr
  }
  }
\source{Yusuf, S. and Peto, R. and Lewis, J. and Collins,
                  R. and Sleight, P. (1984),"Beta blockade during and after myocardial
                  infarction: an overview of the randomized
		  trials",\emph{Progress in Cardiovascular Diseases},
		  \bold{27}, 335--371.
		}
\references{}	
\keyword{datasets}