\name{faults}
\docType{data}
\alias{faults}
\title{Faults in rolls of material}
\description{
  Bissell gives the numbers of yarn breaks observed in a roll
  of fabric whilst a textile process was running, as well as the length of the roll of fabric.
}
\usage{data(faults)}
\format{A data.frame of 32 obs. of 2 variables:
  \tabular{rll}{
  [,1] \tab l \tab integer, roll length (m)\cr
  [,2] \tab n \tab integer, number of faults\cr
}
}
\source{Bissell, A. F. (1972),"A negative binomial model with varying
  elemental sizes", \emph{Biometrika}, \bold{59}, 435--441.
}
\note{See pages 269 and 474 of SMIR}
\keyword{datasets}