% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novel.R
\name{hairpinTestSet}
\alias{hairpinTestSet}
\title{hairpinTestSet}
\usage{
hairpinTestSet(refFileName, dataLst)
}
\arguments{
\item{refFileName}{Full directory and file name of ther reference.}

\item{dataLst}{If or not to apply sample weight.}
}
\value{
Outputs a \code{.fasta} file that can be used as the training set for novel miRNA discovery.
}
\description{
Produce a \code{.fasta} file as the test hairpin data set from the raw
}
\details{
Working with large reference file might result in long running time or system freezing, depending on the hardware configureation (mainly RAM). It is recommanded to build an index for the reference file prior to this operation when the file is large (multi-GB).
}
\examples{
\dontrun{
hairpinTrainset(refFileName = "~/OneDrive/Storey lab/current_work/miRNA_pipeline/reference/hairpin.fa", rawdataLst) # produce the hairpin training set
}
}

