% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novel.R
\name{hairpinTraining}
\alias{hairpinTraining}
\title{hairpinTraining}
\usage{
hairpinTraining(refFileName, dataLst)
}
\arguments{
\item{refFileName}{File name of ther reference.}

\item{wgt}{If or not to apply sample weight. Default is \code{FALSE}.}
}
\value{
Outputs a \code{.fasta} file that can be used as the training set for novel miRNA discovery.
}
\description{
Produce a \code{.fasta} file as the training hairpin data set from the raw
}
\examples{
\dontrun{
hairpinTrainset(refFileName = "~/OneDrive/Storey lab/current_work/miRNA_pipeline/reference/hairpin.fa", rawdataLst) # produce the hairpin training set
}
}

