\name{insult}
\docType{data}
\alias{insult}
\title{Effects of unprovoked verbal attack}
\description{
  Ten male and nine female subjects were asked to fill out a questionnaire which mixed
innocuous questions with questions attempting to assess the subject's
self-reported hostility.  A hostility score for each individual was
calculated from these responses.  After completing the questionnaire,
the subjects were then left waiting for a long time, and were
subjected to insults and verbal abuse by the experimenter when the
questionnaire was eventually collected.  All subjects were told that
they had filled out the questionnaire incorrectly, and were instructed
to fill it out again.  A second hostility score was then calculated
from these later responses.
}
\usage{data(insult)}
\format{A data.frame of 19 obs. of 3 variables:
  \tabular{rll}{
 [,1]\tab hbefore\tab integer, hostility score before verbal attack\cr
 [,2]\tab hafter \tab integer,  hostility score after verbal attack\cr
 [,3]\tab sex    \tab Factor w/ 2 levels "female","male"\cr
  }
  }
\source{Erickson, B. H. and Nosanchuk, T. A., (1979). \emph{Understanding Data},
  Milton Keynes, UK, Open University Press, UK.
}
\note{See pages 5 and 17 of SMIR}
\keyword{datasets}