% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conserved.R
\name{mirFit}
\alias{mirFit}
\title{mirFit}
\usage{
mirFit(dfm, fileName, wgt = FALSE)
}
\arguments{
\item{dfm}{Input dataframe.}

\item{fileName}{output file name. Be sure to use quatation}

\item{wgt}{If or not to apply sample weight. Default is \code{FALSE}.}
}
\value{
Outputs a list with limma eBayes fitting, QC plots and a volcano distribution
}
\description{
data Voom normalization for plotting
}
\examples{
\dontrun{
fitRNA <- mirFit(simDfm, wgt = TRUE)
}
}

