% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conserved-Jing’s MacBook Pro.R, R/conserved.R
\name{mirNrm}
\alias{mirNrm}
\title{mirNrm}
\usage{
mirNrm(dfm, wgt = FALSE)

mirNrm(dfm, wgt = FALSE)
}
\arguments{
\item{dfm}{Input dataframe.}

\item{wgt}{If or not to apply sample weight. Default is \code{FALSE}.}

\item{dfm}{Input dataframe.}

\item{wgt}{If or not to apply sample weight. Default is \code{FALSE}.}
}
\value{
Outputs a dataframe with voom normalization

Outputs a dataframe with voom normalization
}
\description{
data Voom normalization for plotting

data Voom normalization for plotting
}
\examples{
\dontrun{
simNrm <- mirnrm(simDfm, wgt = TRUE)
}
\dontrun{
simNrm <- mirnrm(simDfm, wgt = TRUE)
}
}

