% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conserved-Jing’s MacBook Pro.R, R/conserved.R
\name{mirProcess}
\alias{mirProcess}
\title{mirProcess}
\usage{
mirProcess(wd = getwd())

mirProcess(wd = getwd())
}
\arguments{
\item{wd}{Working directory where all the read count \code{.txt} files are stored. Default is the current working directory.}

\item{wd}{Working directory where all the read count \code{.txt} files are stored. Default is the current working directory.}
}
\value{
Outputs a list with merged read counts from mutliple files, with annotation.

Outputs a list with merged read counts from mutliple files, with annotation.
}
\description{
data pre-processing for miRNA-seq read count files. This function only runs under Unix or Unix-like operating systems. See \code{\link{mirProcessML}}.

data pre-processing for miRNA-seq read count files. This function only runs under Unix or Unix-like operating systems. See \code{\link{mirProcessML}}.
}
\details{
Make sure to follow the file name naming convention for the read count files: \code{ID_database_targettype.txt}

Make sure to follow the file name naming convention for the read count files: \code{ID_database_targettype.txt}
}
\examples{
\dontrun{
readcountMerged <- mirProcess()
}
\dontrun{
readcountMerged <- mirProcess()
}
}

