% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novel.R
\name{mirProcessML}
\alias{mirProcessML}
\title{mirProcessML}
\usage{
mirProcessML(wd = getwd(), setType = "training")
}
\arguments{
\item{wd}{Working directory where all the read count \code{.txt} files are stored. Default is the current working directory.}

\item{setType}{Type of output data set, training or test set. Options are \code{"training"} and \code{"test"}. Default is \code{"training"}.}
}
\value{
Outputs a list with merged read counts from mutliple files, with annotation. No merging inlcuded: see \code{\link{mirProcess}}.
}
\description{
data pre-processing for miRNA-seq read count files specific for machine learning (ML). This function only runs under unix or unix-like operating systems. see \code{\link{mirProcess}}.
}
\details{
Make sure to follow the fie name naming convention for the read count files: \code{ID_database_targettype.txt}
}
\examples{
\dontrun{
readcountML <- mirProcessML()
}
}

