\name{Rsq}
\alias{Rsq}
\title{Coefficient of determination, R-squared}
\description{
Calculates the coefficient of determination for any model of class 'lm'.
}
\usage{
Rsq(model)
}
\arguments{
  \item{model}{a model object list with argument \code{y} and method \code{fitted}}
}
\seealso{\code{cor}}
