\name{stackloss}
\docType{data}
\alias{stackloss}
\title{The Brownlee stackloss data}
\description{
  The data consist of 21 observations on stack-loss
 (the loss of acid through the stack) in a chemical plant for
the conversion of ammonia to nitric acid, with three explanatory
variables: air flow(\code{x1}), cooling water inlet
temperature(\code{x2})  and acid concentration(\code{x3}).
}
\usage{stackloss}
\format{A data.frame of 21 obs. of 4 variables:
  \tabular{rll}{
    [,1] \tab  y \tab integer, loss of acid through the stack \cr
    [,2] \tab  x1 \tab  integer,  air flow\cr
    [,3] \tab  x2 \tab  integer, cooling water inlet temperature \cr
    [,4] \tab  x3 \tab  integer,  acid concentration\cr
}
}
\source{Lange, K L and Little, R J A and Taylor, J M G, (1989). "Robust
  statistical modeling using the $t$ distribution", \emph{J Amer Statist
  Assoc},\bold{84}, 881--896.
 }
\references{}
\keyword{datasets}