\name{toxoplas}
\docType{data}
\alias{toxoplas}
\title{Toxoplasmosis in El Salvador}
\description{
  The file shows the number of men tested and the
number with a positive test for toxoplasmosis in 34 cities in El Slavador,
together with the annual rainfall in metres.
}
\usage{data(toxoplas)}
\format{A data.frame of 34 obs. of 3 variables:
  \tabular{rll}{
 [,1]\tab y \tab integer, number of men with a positive test\cr
 [,2] \tab n \tab  integer, number of men tested \cr
 [,3]\tab  x \tab numeric, annual rainfall in metres, \cr
}
}
\source{Efron, B., (1986), Double exponential families and their use in
                  generalized linear regression, \emph{J Amer Statist
		  Assoc}, \bold{81}, 709--721.
 }
\note{See p.484 in SMIR}
\keyword{datasets}