\name{trees}
\docType{data}
\alias{trees}
\title{The Minitab cherry tree data}
\description{
  The volume of usable
wood \code{v} in cubic feet (1 foot \ = 30.48 cm) is given for each of
a sample of 31 black cherry trees, and the height \code{h} in feet and
the diameter \code{d} in inches (1 inch = 2.54 cm) at a height 4.5 feet
above the ground.
}
\format{A data.frame of 31 obs. of 3 variables:
  \tabular{rll}{
  [,] \tab  d\tab numeric, diameter in inches \cr
 [,] \tab  h\tab integer, height in feet\cr
 [,] \tab  v\tab numeric volume of usable wood,
  }
}
\source{Ryan, T. and Joiner, B. and Ryan, B., (1976). \emph{Minitab
    students handbook}, Duxbury Press, North Scituate, Mass
  }
\usage{trees}
\references{}
\keyword{datasets}