\name{vaso}
\docType{data}
\alias{vaso}
\title{Gilliatt's vaso-constriction data}
\description{
  These data were obtained in a carefully controlled
study of the effect of the rate and volume  of air
inspired by human subjects on the occurrence  or
non-occurrence  of a transient vasoconstriction response in the skin of
the fingers.
}
\format{A data.frame of 39 obs. of 4 variables:
  \tabular{rll}{
    [,1] \tab  Subject\tab Factor w/ 3 levels "1","2","3", ...\cr
    [,2] \tab  Volume \tab numeric, volume of air expired\cr
    [,3] \tab  Rate   \tab numeric, rate of air expired\cr
    [,4] \tab  Y \tab  integer, transient vasoconstriction response, 1=yes, 0=no
}
}
\source{
  Finney, D.J. (1947)., "The estimation from individual records of the
                  relationship between dose and quantal
		  response", \emph{Biometrika}, \bold{34}, 320--34.}
\references{Gilliatt, R.W. (1948). "Vaso-constriction in the finger
  after deep inspiration", \emph{J. Physiol.}, \bold{107}, 76--88.}
\keyword{datasets}