\name{woolson}
\alias{woolson}
\docType{data}
\title{Woolson and Clarke's Obesity Study}
\description{
This is a subset of the Obesity dataset. Binary indicators of obesity on
1014 children who were 7-9 years old in 1977, and were followed up in
1979 and 1981. Children were classified as obese their weights were more
than 210\% of the population median weight for their gender and height.
}
\usage{data(woolson)}
\format{
  A data frame with 48 observations on the following 4 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
    \item{\code{sex}}{a numeric vector}
    \item{\code{age}}{a numeric vector}
  }
}
\details{

}
\source{
Woolson, R.F. and Clark, W.R.(1984). Analysis of categorical incomplete longitudinal
data. \emph{JRSS A}. \bold{147}, 87--99.
}
\references{
}
\examples{
data(woolson)
}
\keyword{datasets}
