% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conserved.R
\name{mirProcess}
\alias{mirProcess}
\title{mirProcess}
\usage{
mirProcess(wd = getwd())
}
\arguments{
\item{wd}{Working directory where all the read count \code{.txt} files are stored. Default is the current working directory.}
}
\value{
Outputs a list with merged read counts from mutliple files, with annotation.
}
\description{
data pre-processing for miRNA-seq read count files. This function only runs under unix or unix-like operating systems.
}
\details{
Make sure to follow the file name naming convention for the read count files: ID_database_targettype.txt
}
\examples{
\dontrun{
readcountMerged <- mirProcess()
}
}

