% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conserved.R
\name{mirProcess}
\alias{mirProcess}
\title{mirProcess}
\usage{
mirProcess(x)
}
\arguments{
\item{x}{TBD}
}
\value{
Outputs a dataframe with merged read counts from mutliple files, with annotation
}
\description{
data pre-processing for miRNA-seq read count files.
}
\examples{
\dontrun{
unfished
}
}

